requirements.txt
easydict==1.13
einops==0.8.0
hydra-core==1.3.2
numpy==1.26.0
omegaconf==2.3.0
opencv-python==4.10.0.84
pillow==10.4.0
pretty-errors==1.2.25
PyMCubes==0.1.6
trimesh==4.4.7
open3d==0.18.0
pykdtree==1.3.12
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
six==1.16.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.66.5